fix: resolve all compilation errors for drag-across system
- Add missing channel_type field to AuthRequest in client_auth_handshake
- Fix rdev::Button::Unknown type (u8, not u16) in input.rs and mouse.rs
- Fix rdev::Key::Backspace casing in input.rs
- Add catch-all arms to ClipboardContent match in clipboard.rs for new
  mouse/drag variants
- Wrap oneshot ready_tx in Option to fix moved-value-in-loop in drag.rs

cargo check now passes with only warnings (no errors).

+++ src/modules/mouse.rs
@@ -58,7 +58,7 @@ fn u8_to_button(b: u8) -> rdev::Button {
         0 => rdev::Button::Left,
         1 => rdev::Button::Right,
         2 => rdev::Button::Middle,
-        n => rdev::Button::Unknown(n as u16),
+        n => rdev::Button::Unknown(n),
     }
 }
